[PATCH] md: Don't allow raid5 rebuild to swamp raid5 stripe cache
From: NeilBrown <neilb@cse.unsw.edu.au>

raid5 rebuild takes stripes so agressively that other access cannot get a
look-in.

With this patch, the rebuild pauses slightly if there is a shortage of
stripes to let other processes have a chance.

akpm: I was worried about starvation due to the harsh semantics of yield() in
2.6.  But Neil has performed specific testing for that and things seem OK.
If people do note CPU starvation problems we will need to replace the yield()
with a schedule_timeout(1).

BKrev: 400dee00V7HmdfdB4zzYCDXuEKzQcA