Merge bk://gkernel.bkbits.net/net-drivers-2.5 into home.osdl.org:/home/torvalds/v2.5/linux
2004/01/24 13:04:35-08:00 torvalds
Merge bk://kernel.bkbits.net/davem/net-2.6
into home.osdl.org:/home/torvalds/v2.5/linux

2004/01/24 10:10:03-08:00 yoshfuji
[IPV6]: Fix several comment spelling errors and typos.

2004/01/24 10:08:59-08:00 jmorris
[IPSEC]: Guard against potentially fatal stack usage for auth_data.

2004/01/24 10:07:26-08:00 bart
[NET]: Return 'unsigned char *' from *skb_pull*() routines.

2004/01/24 09:59:08-08:00 dlstevens
[MULTICAST]: multicast loop with include filters fix

When sending a multicast and using looping back a copy to the
local machine, the interface filter checks can be done before the
source address is specified. For an INCLUDE filter, this won't match
the allowed sources and the packets won't be delivered locally,
even when the ultimate source address chosen is in the allowed list.

The patch below fixes the filter checks for both IGMPv3 and MLDv2
to only apply when a source address is available.

Thanks to Steven Hessing for reporting the problem and providing
a test case for reproducing it.

BKrev: 4012eaf8p7nNQ7Xmv4IUBZGU7YUI0A