Merge bk://kernel.bkbits.net/gregkh/linux/i2c-2.6 into home.osdl.org:/home/torvalds/v2.5/linux
2004/01/27 14:50:36-08:00 greg
Merge kroah.com:/home/greg/linux/BK/bleed-2.5
into kroah.com:/home/greg/linux/BK/i2c-2.6

2004/01/27 14:46:05-08:00 greg
[PATCH] I2C: remove printk() calls in lm85, and clean up debug logic.

2004/01/27 14:38:29-08:00 khali
[PATCH] I2C: Bring lm75 and lm78 in compliance with sysfs naming conventions

Here is a patch that brings the lm75 and lm78 drivers in compliance with
sysfs naming conventions. The drivers as found in existing 2.6 kernels
do not have a digit appended to the temperature-related files names as
the sysfs naming conversion recommends (obviously because they each have
a single temperature channel). As a result, libsensors won't find the
files.

It was discussed on the list wether a '1' should be appended in this
case, and our conclusion was that it would be better to do so because it
helps automatic processing of the sysfs exported files. Please apply if
you agree with this.

2004/01/26 16:58:46-08:00 khali
[PATCH] I2C: Add ADM1025EB support to i2c-parport

The following patch adds support for the ADM1025 evaluation board to the
i2c-parport (and i2c-parport-light) driver(s). In fact, it happens that
it was already supported as an ADM1032 evaluation board, so it is just a
matter of documenting it correctly.

2004/01/23 17:14:52-08:00 khali
[PATCH] I2C: Fix bus reset in i2c-philips-par

This patch fixes the bus reset in i2c-philips-par when it is loaded with
type!=0. For now, the reset is always made as is type==0. I guess that
this driver will be abandoned in a while, but it probably doesn't hurt
to fix that.

2004/01/23 17:14:38-08:00 khali
[PATCH] I2C: undo documentation change

Undo a recent change to the i2c documentation. The change belongs to
2.7.

2004/01/23 17:14:22-08:00 mhoffman
[PATCH] I2C: i2c-piix4.c bugfix

This patch fixes a "Trying to release non-existent resource" error that
occurs during rmmod when the device isn't actually present.  It includes
some other cleanups too: error paths, whitespace, magic numbers, __devinit.

BKrev: 4016f86c2GeVT35AgcgX5Hmmmd1vUw