[PATCH] RAID-6: x86-64 crash workaround
From: "H. Peter Anvin" <hpa@zytor.com>

Apparently, on x86-64, the stack isn't always aligned properly (16 bytes)
in the kernel at the moment.  This causes the RAID-6 code to crash the
system.  This patch is a workaround for that; the right thing is to muck
with the assembly entrypoints to enforce proper stack alignment.  However,
that's not anything I feel comfortable doing in an evening, especially
since I don't have a machine on which I can test the resulting kernels.

BKrev: 401beea5A75mkM58R8jCHxDb8VH-dw