[PATCH] lock_cpu_hotplug only if CONFIG_CPU_HOTPLUG
From: Rusty Russell <rusty@rustcorp.com.au>

The cpucontrol mutex is not required when no cpus can go up and down.
Andrew wrote a wrapper for it to avoid #ifdefs, this expands that to only
be defined for CONFIG_HOTPLUG_CPU, and uses it everywhere.

The only downside is that the cpucontrol lock was overloaded by my recent
patch to net/core/flow.c to protect it from reentrance, so this
reintroduces the local flow_flush_sem.  This code isn't speed critical, so
taking two locks when CONFIG_HOTPLUG_CPU=y is not really an issue.

BKrev: 40213a961g-Vg2sNA1lWn0fitapxsA