[PATCH] Fix more gcc 3.4 warnings
From: Andi Kleen <ak@muc.de>

Just many more warning fixes for a gcc 3.4 snapshot.

It warns for a lot of things now, e.g.  for ?: and ({ ...  }) and casts as
lvalues.  And for functions marked inline in headers, but no body.

Actually there are more warnings, i stopped fixing at some point.  Some of
the warnings seem to be dubious (e.g.  the binfmt_elf.c one, which looks
more like a compiler bug to me)

I also fixed the _exit() prototype to be void because gcc was complaining
about this.

BKrev: 40213abbriK51_p8zQ8xtPf-MePxRA