[PATCH] initialise cpu_vm_mask in init_mm
From: Anton Blanchard <anton@samba.org>

Some architectures use cpu_vm_mask to optimise TLB flushes.  On ppc64 we
are now using a common flush infrastructure that handles both userspace and
kernelspace (vmalloc) pages.  In order to avoid triggering this
optimisation we need to mark the init mm as having scheduled on all cpus.

Things currently work by luck (we check for the cpu only having run on the
local cpu, and the field is initialised to 0), but it would be safer to
initialise it CPU_MASK_ALL.

BKrev: 40213d57MXLnGhomUG6QkLnb88HxVw