[PATCH] md: Move the test in preferred_minor to where it is used.
From: NeilBrown <neilb@cse.unsw.edu.au>

A RAID superblock can indicate which minor number the array should be
assembled under.  As this is only meaningful when doing auto-start, we move
the test for it being in the valid range to the place where auto-start
happens.  When an array is started any other way, it doesn't matter what
value is here.

BKrev: 40213d7bvE3p1HGX8WtsCvC47hIs2g