[PATCH] CDROMREADAUDIO frames
From: Jens Axboe <axboe@suse.de>

2.6 imposes a 64 frame limit where 2.4 does not (just relies on kmalloc()
failing and limiting frames from that).  That breaks at least on guys app.
With MSF adressing, it's much simpler to be able to ask for a full second
at the time, so I think we should just allow that.  So bump the limit from
64 to CD_FRAMES (which is 75).

BKrev: 40213eebmxQTy1JYPSY4s_z9qcqbiw