[PATCH] Fix x86-64 boot problem
From: Andi Kleen <ak@muc.de>

Fix a bug introduced with the last merge that prevented booting with
CONFIG_DEBUG_INFO on on x86-64.  It would corrupt registers in interrupts.
This has hit a few people, so I would consider it as a critical fix.

BKrev: 40213f60coWPQBf6jBmxWRZQQ8pNcw