[PATCH] Fix ptrace in the vsyscall dso area
From: Roland McGrath <roland@redhat.com>

The #include is the part of this patch that matters, so the #ifdef below
works.

The rest of the patch removes gratuitous duplication due to some strange
aversion to concision in the presence of #ifdef, the kind that is all too
common, utterly pointless, and error prone.

BKrev: 40213fd735SjxKnZ5OQdK-6uMlHtIw