[PATCH] paride cleanup and fixes (2/25)
Preparation to crapectomy - expanded include of pseudo.h, split
the part of pd_do_request() into pd_do_request1(), leaving in pd_do_request()
only the check that FSM is not running and call of pd_do_request1().
Added locking analysis - it's in drivers/block/paride/Transition-notes
and it will be used on the next step.  Even though the subsequent steps are
small, they rely on properties of the damn thing and proof of these properties
is hairy.  Amazing perversions people manage to produce when writing a trivial
FSM - the code is very convoluted for no good reason...

BKrev: 402276e7Obtp0gDZpaNk3BJl2mRk5A