[PATCH] cciss: PCI BAR sizing fix
From: <mikem@beardog.cca.cpqcorp.net>

This patch eliminates the bad assumption that all of our PCI BARs will always
be 32-bits.  Tested against the 2.6.2 kernel.

This is required to support the Pinnacles architecture.  It is already in the
2.4 tree.

(This needs to be converted to use pci_request_regions())

BKrev: 4026750bbFEFtw1BGTeuYNpD2ZIrFA