[PATCH] cciss: disble prefetching in ASIC
From: mikem@beardog.cca.cpqcorp.net

This patch addresses a bug in the ASIC on the 6400 series controllers.  When
prefetching from host memory we grab an extra 750 or so bytes of data.  If
this occurs on a memory boundary the machine will MCA.  This bug affects IPF
and Alpha based platforms.  It is not known to be a problem on x86.

Prefetch will be disabled via the f/w.  We need to enable it on x86 to
address a _big_ performance hit on RAID 1 operations.

It is in the 2.4 tree.

BKrev: 40267530CPD7GHGTyXuXEyCUL6u-Sw