Merge bk://kernel.bkbits.net/gregkh/linux/pci-2.6 into home.osdl.org:/home/torvalds/v2.5/linux
2004/02/09 14:05:33-08:00 greg
dmapool: fix up list_for_each() calls to list_for_each_entry()

Now this should get that Rusty^Wmonkey off my back...

2004/02/09 13:25:39-08:00 greg
Merge kroah.com:/home/greg/linux/BK/bleed-2.6
into kroah.com:/home/greg/linux/BK/pci-2.6

2004/02/09 11:04:55-08:00 greg
PCI: move pci_msi.h out of include/linux to drivers/pci where it belongs.

2004/02/09 11:03:22-08:00 greg
PCI: remove stupid MSI debugging code that was never used.

2004/02/06 14:23:33-08:00 eike-hotplug
[PATCH] PCI Hotplug: Convert error paths in ibmphp to use goto

This patch converts the error paths in several functions in ibmphp to use
goto to the end of the function to kill code duplication.

It also kills some memsets of pointer members of a struct where the struct
itself is freed directly after.

This one has 2 modification from the first version: the if is back as
discussed and one extra long line (>> 150 chars) is wrapped now.

2004/02/04 17:23:50-08:00 eike-hotplug
[PATCH] PCI Hotplug: mark functions __init/__exit in acpiphp

These functions are only called from places marked __init or __exit, so
we can mark them also.

2004/02/04 14:03:31-08:00 greg
[PATCH] dmapool: Remove sentance in documentation that is now false on 2.6

2004/02/04 10:03:35-08:00 colpatch
[PATCH] PCI: fix "pcibus_class" Device Class, release function

John Rose wrote:
> The function release_pcibus_dev() in probe.c defines the release procedure for
> device class pcibus_class.  I want to suggest that this function be scrapped :)
>
> This release function is called in the code path of class_device_unregister().
> The pcibus_class devices aren't currently unregistered anywhere, from what I
> can tell, so this release function is currently unused.  The runtime removal of
> PCI buses from logical partitions on PPC64 requires the unregistration of these
> class devices.  The natural place to do this IMHO is in pci_remove_bus_device()
> in remove.c.

You're right that the class device isn't currently unregistered, and
that was an oversight in the patch I originally sent.  Attatched is a
patch that remedies that situation.  pci_remove_bus_device() *is* the
natural place to unregister the class_dev, and that's just what the
patch does.


> The problem is that this calls pci_destroy_dev(), which calls put() on the same
> "bridge" device that the release function does.  This should only be done once
> in the course of removing a pci_bus, and I doubt that we want to change
> pci_destroy_dev().   The kfree() of the pci_bus struct is also done in both
> pci_remove_bus_device() and release_pcibus_dev().

Yep.  The patch pulls the kfree() out of pci_remove_bus_device() and
calls class_device_unregister() in it's place.  This will put() the
bridge device and free the pci_bus as needed.  put() does need to be
called twice because the bridge device is get()'d twice: once when the
device is registered and once when it's bus device grabs a reference to it.

2004/02/03 16:54:54-08:00 eike-hotplug
[PATCH] PCI Hotplug: make ibm_unconfigure_device void

ibm_unconfigure_device always returns 0, so we do not need to check for a
return value != 0 and can kill the warning. And if the return value is always
the same we don't need a return value. Also some whitespace fixing. And this
time not line wrapped. Really. Absolutely sure. Checked twice.

2004/02/03 16:49:40-08:00 eike-hotplug
[PATCH] PCI Hotplug: kill hpcrc from several functions in ibmphp_core.c

This patch kills the variable hpcrc from many places. It is used as a
temporary storage for the return code before this is copied to rc which
contains the "real" return code. There are some checks if to copy or not
but at the end they will always have the same value, so we can directly
put the return code in rc and kill hpcrc.

2004/02/03 16:49:23-08:00 eike-hotplug
[PATCH] PCI Hotplug: very small optimisations for ibmphp_pci.c

2004/02/03 16:49:05-08:00 eike-hotplug
[PATCH] PCI Hotplug: Coding style fixes for drivers/pci/hotplug.c

2004/02/03 16:48:38-08:00 eike-hotplug
[PATCH] PCI: avoid two returns directly after each other in pcidriver.c

This avoids a return direct before the final return of a function. Better
only modify the return code and fall through to the final return.

2004/02/03 16:22:27-08:00 david-b
[PATCH] PCI: dma_pool fixups

2004/02/02 12:19:50-08:00 dlsy
[PATCH] PCI: Patch to get cpqphp working with IOAPIC

Here is a patch for to get cpqphp working with IOAPIC.
My earlier statement that a kernel patch is not needed for 2.6 is
true only when ACPI is enabled.  A similar patch is needed in
pcibios_enable_irq() for 2.4 kernel and I will send it out
later.

The fix is in pirq_enable_irq().  This function is called indirectly
by pci_enable_device().  For device present during boot up, it should
get the proper dev->irq for pcibios_fixup_irqs() has been called to
get the dev->irq from MP table. If the value is still zero, then
this is properly caused by "buggy MP table".  For hot-plug device,
its dev->irq is 0 when the pci_enable_device() is called for it hasn't
gone through the fixup.  Therefore, the code (similiar to the code
in pcibios_fixup_irqs) is needed here.

2004/02/02 11:50:52-08:00 eike-hotplug
[PATCH] PCI Hotplug: Kill useless instructions from ibmphp_core.c

This kills the explicit setting of rc to -ENODEV in 2 places where it is
not necessary because it will have this value on this path anyway.

2004/02/02 11:42:58-08:00 eike-hotplug
[PATCH] PCI Hotplug: Kill spaces before \n in ibmphp*

Don't use a space directly before '\n' in error/debug messages. This adds
extra code size, causes unneeded line breaks in xterms, noone else does it
and it's just superfluos and ugly.

This also adds a missing ")" in a comment.

2004/02/02 11:42:37-08:00 eike-hotplug
[PATCH] PCI Hotplug: Whitespace fixes for acpiphp

This are some whitespace fixes for acpiphp, in most cases killing spaces
before the opening brace of function declarations.

2004/02/02 11:29:59-08:00 greg
[PATCH] dmapool: use dev_err() whenever we can to get the better information in it.

2004/02/02 11:05:17-08:00 dsaxena
[PATCH] PCI: Replace pci_pool with generic dma_pool, part 2

include/linux changes:

- Add dma_pools member to struct device
- Add include/linux/dmapool.h
- Remove pools memober from struct pci_dev
- Replace pci_pool_* functions with macros that map to dma_pool_* functions

2004/02/02 11:04:46-08:00 dsaxena
[PATCH] PCI: Replace pci_pool with generic dma_pool

- Move drivers/pci/pool.c to drivers/base/pool.c
- Initialize struct device.dma_pools in device_initialize()
- Remove initialization of struct pci_dev.pools from pci_setup_device()

2004/01/30 16:35:12-08:00 eike-hotplug
[PATCH] PCI Hotplug: coding style for cpqphp_pci.c

here are some coding style fixes.

2004/01/30 16:34:48-08:00 ambx1
[PATCH] PCI: Remove uneeded resource structures from pci_dev

The following patch remove irq_resource and dma_resource from pci_dev.  It
appears that the serial pci driver depends on irq_resource, however, it may be
broken portions of an old quirk.  I attempted to maintain the existing behavior
while removing irq_resource.  I changed FL_IRQRESOURCE to FL_NOIRQ.  Russell,
could you provide any comments?  irq_resource and dma_resource are most likely
remnants from when pci_dev was shared with pnp.

2004/01/30 16:31:28-08:00 greg
[PATCH] PCI: add back some pci.ids entries that got deleted in the last big update.

BKrev: 40281a6cx5qSM0PE2ITd9kV_xyhoLA