Merge bk://kernel.bkbits.net/gregkh/linux/i2c-2.6 into home.osdl.org:/home/torvalds/v2.5/linux
2004/02/09 14:09:50-08:00 greg
Merge kroah.com:/home/greg/linux/BK/bleed-2.6
into kroah.com:/home/greg/linux/BK/i2c-2.6

2004/02/09 13:18:09-08:00 khali
[PATCH] I2C: add new chip driver: fscher

This is a new ported driver, fscher, which supports the FSC Hermes chip.
The original driver was written by Reinhard Nissl, who also ported it to
Linux 2.6, as discussed on the lm_sensors mailing list during the last
two weeks. I reviewed the code and we made the necessary changes, so
that what we have now looks good to me. Please apply on top of your i2c
patches stack.

2004/02/05 17:04:24-08:00 khali
[PATCH] I2C: Update I2C maintainers entry

I propose the following update of the "I2C AND SENSORS DRIVERS" entry in
MAINTAINERS:

* Remove Frodo Looijaard. He isn't active in the project anymore, and
forwards everythings he receives to the mailing-list. More work for
everyone, and increased latency.

* Move your name to the top. This is more realistic since you are the
real maintainer of the i2c subsystem in 2.6.

* Reindent. Spaces were used instead of tabs.

2004/02/03 17:01:31-08:00 greg
[PATCH] I2C: fix oops when CONFIG_I2C_DEBUG_CORE is enabled and the bttv driver is loaded.

2004/02/02 15:04:34-08:00 khali
[PATCH] I2C: add new driver, gl518sm

Here comes a new i2c chip driver for Linux 2.6: gl518sm, ported from our
CVS repository by Hong-Gunn Chew. I have reviewed Gunn's code and we
made a few cleanups and improvements over the original driver. See our
complete exchange on the sensors mailing-list for details.

2004/02/02 14:52:07-08:00 khali
[PATCH] I2C: Handle read errors in w83l785ts

Here is a patch that adds proper read error handling in the w83l785ts
driver. This is needed for this driver because on many Asus boards the
BIOS or something accesses the chip in our back and causes collisions or
something similar that causes many read errors. For now, these errors
make the driver return temperature values of -1 from times to times.

I have been working with James Bolt on this. See the thread on the
lm_sensors mailing list for details. The patch is fairly well tested. It
is against 2.6.2-rc3 + your current stack of patches as I know it.

The "retry until it works" method is the best I could think of, since we
have no information from Asus and I don't expect to get any. If it still
doesn't work after an arbitrary number (5) of tries it returns the
previously known value and generates an error in the logs. James' tests
showed that it will probably never happen though (highest retry count
was 3 and happened once out of 3000 reads) unless we lower the arbitrary
number (but we don't want to, do we?)

I inserted incremental delays as reads fail, I felt like it should help
avoid collisions with whatever-is-bugging-us. Seems to work OK, but it's
not perfect (since we sometimes need 3 retries) and I didn't test with a
different policy (no delay, constant delay or different increment).

2004/02/02 14:51:45-08:00 khali
[PATCH] I2C: Bring w83l785ts in compliance with sysfs naming conventions

Here is a patch that brings the w83l785ts driver in compliance with
sysfs naming conventions. This is pretty much the same problem and
solution that occured very recently with the lm75 and lm78 drivers.

The patch was tested to work fine by James Bolt.

2004/01/30 16:56:36-08:00 ender
[PATCH] I2C: fix typos

Following patch fixes two typos and a missing full stop.

2004/01/30 16:40:31-08:00 geert
[PATCH] I2C: add Hydra i2c bus driver

Here's a new version, incorporating these comments, and making a few more
changes:
  - Use struct definition in <asm/hydra.h> instead of #defined offset
  - Remove flushes are register writes, they are no longer needed
  - Use pci_resource_start() instead of dev->resource[].start
  - ioremap() the whole resource instead of the first 256 bytes
  - Check for errors returned by ioremap() and i2c_bit_add_bus()
  - Add resource management
  - Call iounmap() in hydra_remove() instead of in i2c_hydra_exit()
  - Let I2C_HYDRA depend on I2C and select I2C_ALGOBIT instead of depending on
    I2C_ALGOBIT

BKrev: 40281ad91aNptqO64JKu5h6Fkxm7_Q