[PATCH] Export OF device path for PCI devices
Here's a new version of the patch according to our discussion.

This allows the platform to add its own platform files to the
sysfs PCI device node through a "pcibios_add_platform_entries()"
call.

I added an empty pcibios_add_platform_entries() inline to all
archs that apparently have PCI and implemented the OF stuff for
ppc and ppc64. The name is still "devspec" for now, I doubt
it will conflict and that's consistent with our use of the
"devspec" name in other places.

BKrev: 402902c0YzaHoL0_1UaAjRiguN-uMQ