[PATCH] back out fbdev sysfs support
This backs out James' sysfs support for fbdev again.  It introduces a
big, race for every driver not converted to framebuffer_{alloc,release}
(that is every driver but Ben's new radeonfb).

I've left in framebuffer_{alloc,release} as stubs so drivers can be
converted to it gradually and once all drivers are done it can be
enabled again.

BKrev: 402e55000UBbvngfi2u7cQo2l89O8g