[PATCH] fix FB_RADEON_I2C dependency
Thus fixes the weird kconfig message "optimize || ?", it's an old debug
check and is triggered by the unusual dependency.  It's not incorrect,
but the solution below is better and it's the same FB_MATROX_I2C already
uses.

I'll send a fix for the kconfig message later.

BKrev: 402e7c43mOqSItR96RbDxcvSihd6qQ